Developing

+++ app/Controllers/EducationController.php
@@ -15,11 +15,11 @@ class EducationController extends Controller
      */
     public function index()
     {
-        $userId = Helper::getIdFromUrl('user');
+        $userId = Helper::getUserIdFromSession();
         $educations = EducationModel::load()->getAllByUserId($userId);
         $user = UserModel::load()->get($userId);
 
-        View::render('educations/index.view',[
+        return View::render('educations/index.view',[
             'educations' => $educations,
             'user' => $user
         ]);
@@ -32,8 +32,7 @@ class EducationController extends Controller
     {   
         return View::render('educations/create.view', [
             'method'    => 'POST',
-            'action'    => '/education/store',
-            'users'     => UserModel::load()->all(),
+            'action'    => '/education/store'
         ]);
     }
 
@@ -43,9 +42,13 @@ class EducationController extends Controller
     public function store()
     {        
         // Save post data in $education var
+        if(!(int)$_POST['end_year']){
+            $_POST['end_year'] = NULL;
+        }
         $education = $_POST;
 
         // Set created_by ID and set the date of creation
+        $education['user_id'] = Helper::getUserIdFromSession();
         $education['created_by'] = Helper::getUserIdFromSession();
         $education['created'] = date('Y-m-d H:i:s');
 
@@ -53,8 +56,7 @@ class EducationController extends Controller
         EducationModel::load()->store($education);
         
         // Return to the user-overview
-        $userId = $education['user_id'];
-        header("Location: /user/$userId/educations");
+        header("Location: /educations");
     }
 
     /**
@@ -65,12 +67,23 @@ class EducationController extends Controller
         $educationId = Helper::getIdFromUrl('education');
         $education = EducationModel::load()->get($educationId);
 
-        return View::render('educations/edit.view', [
-            'method'    => 'POST',
-            'action'    => '/education/' . $educationId . '/update',
-            'education' => $education,
-            'users'     => UserModel::load()->all(),            
-        ]);
+        if ($education->user_id == Helper::getUserIdFromSession())
+        {
+            return View::render('educations/edit.view', [
+                'method'    => 'POST',
+                'action'    => '/education/' . $educationId . '/update',
+                'education' => $education,
+                'users'     => UserModel::load()->all(),            
+            ]);
+        }
+        else
+        {
+            return View::render('errors/403.view', [
+                'message' => 'you shouldnt edit someone elses education'
+            ]);
+        }
+
+        
     }
 
     /**
@@ -79,9 +92,13 @@ class EducationController extends Controller
     public function update()
     {
         $educationId = Helper::getIdFromUrl('education');
+        if(!(int)$_POST['end_year']){
+            $_POST['end_year'] = NULL;
+        }
         $education = $_POST;
 
         // Set updated_by ID and set the date of updating
+        $education['user_id'] = Helper::getUserIdFromSession();
         $education['updated_by'] = Helper::getUserIdFromSession();
         $education['updated'] = date('Y-m-d H:i:s');
 
@@ -89,8 +106,7 @@ class EducationController extends Controller
         EducationModel::load()->update($education, $educationId);
 
         // Return to the user-overview
-        $userId = $education['user_id'];
-        header("Location: /user/$userId/educations");
+        header("Location: /educations");
     }
 
     /**
@@ -107,9 +123,21 @@ class EducationController extends Controller
     public function destroy()
     {
         $educationId = Helper::getIdFromUrl('education');
-        $userId = EducationModel::load()->get($educationId)->user_id;
-        EducationModel::load()->destroy($educationId);
-        header("Location: /user/$userId/educations");
+        $education = EducationModel::load()->get($educationId);
+
+        if ($education->user_id == Helper::getUserIdFromSession())
+        {
+            EducationModel::load()->destroy($educationId);
+        }
+        else
+        {
+            return view::render('errors/views', [
+                'message'   => 'You can only delete your own education'
+            ]);
+        }
+        
+
+        header("Location: /educations");
     }
 
 }
